requirements/dev.txt
-r base.txt

django-debug-toolbar
responses
mockredispy
flake8
coverage
pytest
pytest-mock
pytest-django
pytest-xdist
